version: Rename VERSION file to prevent header name collisions
Somewhere in the macOS build, there's an #include for <version>. Because the
root source directory is added to the include path during build, that include
picks up this VERSION file, which makes the build fail since it doesn't contain
valid C code. Renaming the file to prevent the name collision allows the build
to proceed as normal.

Signed-off-by: Ruby Iris Juric <ruby@srxl.me>